Add a note message to suggest user to add vue.config,js file in the root manually if not present.
With Vue CLI 3 vue.config.js file is removed from the root of the project. For beginners like me, it is a bit confusing whether the vue.config.js file is hidden or placed somewhere else this can create unnecessary frustration.  Now with this note user will know what to do if they didn't find the vue.config.js file.